+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 
 # Change Log
 
+## Version 2.6.1
+
+* [FIXED] - #263 - Added server URLs to connect exception (not to auth exception)
+* [FIXED] - #262 - Added @deprecated as needed
+* [FIXED] - #261 - Added a static credentials implementation that uses char arrays
+* [FIXED] - #260 - Moved to nats-server from gnatsd for testing
+* [FIXED/CHANGED] - #259 - Double authentication errors from a server during reconnect attempts will result in the connection being closed.
+* [FIXED] - #257 - Added connection method to messages that come from subscriptions, dispatchers and requests
+* [FIXED] - #243 - Added check for whitespace in subjects and queue names
+* [FIXED] - Improved a couple flaky tests
+
 ## Version 2.6.0
 
 * [FIXED] - cleaned up use of "chain" instead of "creds"